+++ src/group_impl.h
@@ -82,9 +82,9 @@ static void secp256k1_ge_set_gej_var(secp256k1_ge_t *r, secp256k1_gej_t *a) {
 }
 
 static void secp256k1_ge_set_all_gej_var(size_t len, secp256k1_ge_t r[len], const secp256k1_gej_t a[len]) {
-    int count = 0;
+    size_t count = 0;
     secp256k1_fe_t az[len];
-    for (int i=0; i<len; i++) {
+    for (size_t i=0; i<len; i++) {
         if (!a[i].infinity) {
             az[count++] = a[i].z;
         }
@@ -94,7 +94,7 @@ static void secp256k1_ge_set_all_gej_var(size_t len, secp256k1_ge_t r[len], cons
     secp256k1_fe_inv_all_var(count, azi, az);
 
     count = 0;
-    for (int i=0; i<len; i++) {
+    for (size_t i=0; i<len; i++) {
         r[i].infinity = a[i].infinity;
         if (!a[i].infinity) {
             secp256k1_fe_t *zi = &azi[count++];
